KeyValue转换器切分方式修改
