+++ parser/logfmt/logfmt_test.go
@@ -1,14 +1,12 @@
 package logfmt
 
 import (
-	"fmt"
 	"testing"
-
-	"github.com/qiniu/logkit/parser"
 
 	"github.com/stretchr/testify/assert"
 
 	"github.com/qiniu/logkit/conf"
+	"github.com/qiniu/logkit/parser"
 	. "github.com/qiniu/logkit/parser/config"
 	"github.com/qiniu/logkit/utils"
 	. "github.com/qiniu/logkit/utils/models"
@@ -31,120 +29,6 @@ func Benchmark_ParseLine(b *testing.B) {
 		m, _ = p.Parse(testData)
 	}
 	bench = m
-}
-
-func Test_parseLine(t *testing.T) {
-	tests := []struct {
-		line       string
-		keepString bool
-		expectData []Data
-		existErr   bool
-		splitter   string
-	}{
-		{
-			expectData: []Data{},
-			existErr:   true,
-			splitter:   "=",
-		},
-		{
-			line: "foo=\"bar\"",
-			expectData: []Data{
-				{"foo": "bar"},
-			},
-			splitter: "=",
-		},
-		{
-			line: "foo=\"bar\"\n",
-			expectData: []Data{
-				{"foo": "bar"},
-			},
-			splitter: "=",
-		},
-		{
-			line: "ts=2018-01-02T03:04:05.123Z lvl=info msg=\"http request\" method=PUT\nduration=1.23 log_id=123456abc",
-			expectData: []Data{
-				{
-					"lvl":      "info",
-					"msg":      "http request",
-					"method":   "PUT",
-					"duration": 1.23,
-					"ts":       "2018-01-02T03:04:05.123Z",
-					"log_id":   "123456abc",
-				},
-			},
-			splitter: "=",
-		},
-		{
-			line: `ts=2018-01-02T03:04:05.123Z lvl=info  method=PUT msg="http request" a=12345678901234567890123456789`,
-			expectData: []Data{
-				{
-					"lvl":    "info",
-					"msg":    "http request",
-					"method": "PUT",
-					"ts":     "2018-01-02T03:04:05.123Z",
-					"a":      1.2345678901234568e+28,
-				},
-			},
-			splitter: "=",
-		},
-		{
-			line:       `ts=2018-01-02T03:04:05.123Z lvl=info  method=PUT msg="http request" a=12345678901234567890123456789`,
-			keepString: true,
-			expectData: []Data{
-				{
-					"lvl":    "info",
-					"msg":    "http request",
-					"method": "PUT",
-					"ts":     "2018-01-02T03:04:05.123Z",
-					"a":      "12345678901234567890123456789",
-				},
-			},
-			splitter: "=",
-		},
-		{
-			line:       `no data.`,
-			expectData: []Data{},
-			existErr:   true,
-			splitter:   "=",
-		},
-		{
-			line:       `foo="" bar=`,
-			expectData: []Data{},
-			existErr:   true,
-			splitter:   "=",
-		},
-		{
-			line: `abc=abc foo="def`,
-			expectData: []Data{
-				{
-					"abc": "abc",
-					"foo": "\"def",
-				},
-			},
-			existErr: false,
-			splitter: "=",
-		},
-		{
-			line:       `"foo=" bar=abc`,
-			expectData: []Data{},
-			existErr:   true,
-			splitter:   "=",
-		},
-	}
-	l := Parser{
-		name: TypeLogfmt,
-	}
-	for _, tt := range tests {
-		l.keepString = tt.keepString
-		l.splitter = tt.splitter
-		got, err := l.parse(tt.line)
-		fmt.Println("got: ", got, " err: ", err)
-		assert.Equal(t, tt.existErr, err != nil)
-		assert.Equal(t, len(tt.expectData), len(got))
-		for i, m := range got {
-			assert.Equal(t, tt.expectData[i], m)
-		}
-	}
 }
 
 func TestParse(t *testing.T) {
@@ -295,7 +179,7 @@ func TestParseWithKeepRawData(t *testing.T) {
 					"lvl":      float64(5),
 					"msg":      "error",
 					"log_id":   "123456abc",
-					"method":"PUT",
+					"method":   "PUT",
 					"duration": 1.23,
 					"raw_data": "ts:2018-01-02T03:04:05.123Z lvl:5 msg:\"error\" log_id:123456abc\nmethod:PUT duration:1.23 log_id:123456abc",
 				},
@@ -331,156 +215,5 @@ func GetParseTestData(line string, size int) []string {
 		}
 		testSlice = append(testSlice, line)
 		totalSize += len(line)
-	}
-}
-
-func Test_splitKV(t *testing.T) {
-
-	tests := []struct {
-		line       string
-		expectData []string
-		existErr   bool
-		splitter   string
-	}{
-		{
-			line: "foo=",
-			expectData: []string{
-				"foo",
-			},
-			existErr: false,
-			splitter: "=",
-		},
-		{
-			line: "=def",
-			expectData: []string{
-				"",
-				"def",
-			},
-			existErr: false,
-			splitter: "=",
-		},
-		{
-			line: "foo=def abc = abc ",
-			expectData: []string{
-				"foo",
-				"def",
-				"abc",
-				"abc",
-
-			},
-			existErr: false,
-			splitter: "=",
-		},
-		{
-			line: "foo\n=def\nabc=a\nbc",
-			expectData: []string{
-				"foo",
-				"def",
-				"abc",
-				"a\nbc",
-			},
-			existErr: false,
-			splitter: "=",
-		},
-		{
-			line: "foo=def \n abc =abc",
-			expectData: []string{
-				"foo",
-				"def",
-				"abc",
-				"abc",
-			},
-			existErr: false,
-			splitter: "=",
-		},
-		{
-			line: "foo=def\n abc=abc",
-			expectData: []string{
-				"foo",
-				"def",
-				"abc",
-				"abc",
-			},
-			existErr: false,
-			splitter: "=",
-		},
-		{
-			line: "foo=def\n test abc=abc",
-			expectData: []string{
-				"foo",
-				"def\n test",
-				"abc",
-				"abc",
-			},
-			existErr: false,
-			splitter: "=",
-		},
-		{
-			line: "time=2018-01-02T03:04:05.123Z \nCST abc=abc",
-			expectData: []string{
-				"time",
-				"2018-01-02T03:04:05.123Z \nCST",
-				"abc",
-				"abc",
-			},
-			existErr: false,
-			splitter: "=",
-		},
-		{
-			line: "foo:de:f ad abc:a:b:c",
-			expectData: []string{
-				"foo",
-				"de:f ad",
-				"abc",
-				"a:b:c",
-			},
-			existErr: false,
-			splitter: ":",
-		},
-		{
-			line: "f:o:o::def",
-			expectData: []string{
-				"f:o:o",
-				"def",
-			},
-			existErr: false,
-			splitter: "::",
-		},
-		{
-			line:       "f:o:o:def",
-			expectData: []string{},
-			existErr:   true,
-			splitter:   "::",
-		},
-		{
-			line:       "f:o:o:\n:def",
-			expectData: nil,
-			existErr:   true,
-			splitter:   "::",
-		},
-		{
-			line: "f:\no::a o:\n:def",
-			expectData: []string{
-				"f:\no",
-				"a o:\n:def",
-			},
-			existErr: false,
-			splitter: "::",
-		},
-		{
-			line:       "f:o:o: \n :def",
-			expectData: []string{},
-			existErr:   true,
-			splitter:   "::",
-		},
-	}
-	for _, tt := range tests {
-		got, err := splitKV(tt.line, tt.splitter)
-		assert.Equal(t, tt.existErr, err != nil)
-		assert.Equal(t, len(tt.expectData), len(got))
-		for i, m := range got {
-			assert.Equal(t, tt.expectData[i], m)
-		}
-
 	}
 }
