+++ parser/syslog/syslog_test.go
@@ -144,7 +144,7 @@ func TestSyslogParser_TimeZone(t *testing.T) {
 	assert.Nil(t, err)
 	lines := []string{
 		`<1>Feb 05 01:02:03 abc system[253]: Listening at 0.0.0.0:3000`,
-		`<34>1 2019-02-04T17:02:03.000Z mymachine.example.com su - ID47`,
+		`<34>1 2020-02-04T17:02:03.000Z mymachine.example.com su - ID47`,
 	}
 	dts, err := p.Parse(lines)
 	assert.Nil(t, err)
@@ -152,6 +152,6 @@ func TestSyslogParser_TimeZone(t *testing.T) {
 	assert.Nil(t, err)
 	dts = append(dts, ndata...)
 	for _, dt := range dts {
-		assert.Equal(t, "2019-02-04 17:02:03 +0000 UTC", dt["timestamp"].(time.Time).String())
+		assert.Equal(t, "2020-02-04 17:02:03 +0000 UTC", dt["timestamp"].(time.Time).String())
 	}
 }
